nexa-analise-avancada-de-imagens-e-texto-com-ai-na-aws/challenges/rekognition_celebrities/pyproject.toml
[project]
name = "rekognition-celebrities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.7",
    "boto3-stubs[rekognition]>=1.36.7",
    "pillow>=11.1.0",
]
